🔨 refactor: Apply rank util and component
