+++ frontend/src/pages/ProfilePage/HistoryBox/HistoryBox.tsx
@@ -4,7 +4,8 @@ import { Grid } from '@/layout/Grid';
 import { Box } from '@/common/Box';
 import { Text } from '@/common/Text';
 import { Avatar } from '@/common/Avatar';
-import { getRank, RankBadge } from '@/pages/RankUtil/RankUtil';
+import { RankBadge } from '@/common/RankBadge';
+import { getRank } from '@/libs/utils/rank';
 import { darken, lighten } from 'polished';
 import { historyMockData, HistoryProps, Player } from './history-mock-data';
 const { histories, histories_fetch } = historyMockData;
