+++ frontend/src/pages/ProfilePage/ProfilePage.tsx
@@ -4,7 +4,8 @@ import { Avatar } from '@/common/Avatar';
 import { InfoBox } from './InfoBox';
 import { AchievementBox } from './AchievementBox';
 import { HistroyBox } from './HistoryBox';
-import { RankBadge, getRank } from '../RankUtil';
+import { RankBadge } from '@/common/RankBadge';
+import { getRank } from '@/libs/utils/rank';
 import { useMemo } from 'react';
 import { ReactComponent as Achievement1 } from '@/svgs/achievment1.svg';
 import { ReactComponent as Achievement2 } from '@/svgs/achievment2.svg';
@@ -25,7 +26,7 @@ export const ProfilePage = () => {
         <Text size="xxl">{nickname}</Text>
       </Grid>
       <Grid container="flex" direction="row" justifyContent="space-between" alignItems="center" columnGap={2}>
-        <InfoBox title="랭크" desc={rank} component={<RankBadge rank={rank} width="100%" />} />
+        <InfoBox title="랭크" desc={rank} component={<RankBadge rank={rank} size="xl" />} />
         <InfoBox title="승률" desc={`${winRate}%`} subDesc={`${winCount}승 ${loseCount}패`} />
         <InfoBox title="경험치" desc={exp.toString()} subDesc="exp" />
       </Grid>
