--- frontend/src/pages/RankUtil/index.ts
@@ -1,1 +1,0 @@
-export * from './RankUtil';
